+++ minimal-examples-lowlevel/secure-streams/minimal-secure-streams-server/ss-client.c
@@ -58,7 +58,7 @@ myss_state(void *userobj, void *sh, lws_ss_constate_t state,
 	myss_t *m = (myss_t *)userobj;
 
 	lwsl_user("%s: %p %s, ord 0x%x\n", __func__, m->ss,
-		  lws_ss_state_name((int)state), (unsigned int)ack);
+		  lws_ss_state_name(state), (unsigned int)ack);
 
 	switch (state) {
 	case LWSSSCS_CREATING:
